Add onerror to XMLHttpRequest
If there is an error, onload will not be called, but we should still
increment processedPairs and draw edges if it's the last pair.

This makes it possible to test in Safari on a local machine again.
When doing so, onerror is called when a blurb is not found, whereas
when testing on a remote server onload is called with a status of 404.

Local testing still doesn't work in Firefox, Chrome, or Opera, but at
least it's something. :/

See issue #13 and 97ccef7f97ca0b53a144090b4475e55c36d2885a