Pull individual character descriptions from Wikipedia
Currently pull ALL characters' descriptions from their respective Wikipedia
articles.  In the future, we may want to pull ASOIAF descriptions from
A Wiki of Ice and Fire or the "List of A Song of Ice and Fire characters"
Wikipedia page.

Currently use each vertex's name as the Wikipedia handle; for some characters
these may not be the same, so we may have to add a wiki_handle variable to
the vertex objects.

See #17