Add PHP script to handle wiki extracts
This is necessary because of the same-origin policy.  A client-side script
can't make requests from websites in another domain, because in doing so it
would send the browser's cookies for that domain, which could be used for
malicious purposes.  See https://en.wikipedia.org/wiki/Same-origin_policy.

To get around this, we access the other domains (in this case the wikis) using
a script on the server.  Then the client-side script (i.e., main.js) is making
a request to the same server it comes from, which does not violate the same-origin
policy, since in this case the browser sends cookies for the site the client-side
script came from, NOT the third-party site it's trying to access.

Fixes #17